+++ code/src/components/NewTodo/NewTodo.js
@@ -2,101 +2,8 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import uniqid from 'uniqid';
 import moment from 'moment';
-import styled from 'styled-components';
-import todos from '../reducers/store';
-
-const Container = styled.div`
-/* grid-column: 1 / 2; 
-grid-row: 1 / 2;  */
-padding: 2em; 
-
-`
-
-const TodoForm = styled.form`
-display: grid; 
-grid-template-columns:  1fr 1fr ; 
-gap: 1em; 
-
-@media (max-width: 678px){
-  display:flex;
-  flex-direction: column; 
-  
-}
- `
-
-const TodoInput = styled.input`
-width: 80%; 
-height: 50%;
-display: inline-block;
-justify-content: center;
-margin-top:0.5em; 
-margin-left:1.5em; 
-grid-column: 1 / 1; 
-border-radius: 15px; 
-font-family: 'Reenie Beanie';
-font-weight:bold; 
-font-size:2em; `
-
-const RadioBtnConatiner = styled.div`
-display: block;
-text-align: center; 
-@media (max-width: 678px){
-  display:flex;
-  flex-direction: column; 
-  
-}
-
-  `
-
-const RadioBtn = styled.input`
-margin-top: 1.5em; 
-margin-right: 10px;
-margin-left: 10px;  
-`
-
-const DueDate = styled.label`
-margin-top: 10px; 
-font-family: 'Reenie Beanie';
-font-weight:bold; 
-font-size:2em; 
-    `
-
-const SubmitBtn = styled.button`
-margin-right: 20px;
-text-align: center; 
-display: inline-block;
-border-radius: 15px;  
-padding: 0.5rem 0; 
-margin: 0.5rem 1rem;
-width: 11rem;
-background-color: #070f4e;
-border: 2px solid white;
-font-weight: bold; 
-color: #f5ebeb; 
-@media (max-width: 678px){
-  text-align: center;  
-  
-}
-
-  `
-
-const TextInput = styled.div`
- `
-
-const SubmittBtnConatiner = styled.div`
-padding-left: 1.5em; 
-margin-top:15px;  `
-
-const HeadingContainer = styled.div`
-display: flex; 
-flex-direction: row;
-justify-content: center; 
-margin-bottom: 10px;   
- `
-const Heading = styled.h1`
-padding: 10px; 
-color: #f5ebeb; 
-background-color: #070f4e;`
+import todos from '../../reducers/store';
+import { Container, TodoForm, TextInput, HeadingContainer, Heading, TodoInput, RadioBtnContainer, DueDate, RadioBtn, SubmittBtnContainer, SubmitBtn } from './NewTodo.style';
 
 const NewTodo = () => {
   const [input, setInput] = useState('')
@@ -105,7 +12,7 @@ const NewTodo = () => {
 
   const today = moment().format('ddd D MMM');
   const tomorrow = moment().add(1, 'days').format('ddd D MMM');
-  const yesterday = moment().add(-1, 'days').format('ddd D MMM')
+  const twodays = moment().add(2, 'days').format('ddd D MMM')
 
   console.log(today, tomorrow)
   const onNewTodo = (event) => {
@@ -139,7 +46,7 @@ const NewTodo = () => {
             onChange={(event) => setInput(event.target.value)}
             required />
         </TextInput>
-        <RadioBtnConatiner>
+        <RadioBtnContainer>
           <HeadingContainer>
             <Heading>DUE DATE</Heading>
           </HeadingContainer>
@@ -162,15 +69,15 @@ const NewTodo = () => {
           <DueDate>
             <RadioBtn
               type="radio"
-              value={yesterday}
+              value={twodays}
               name="date"
               onChange={(event) => SetDueDate(event.target.value)} />
-            {yesterday}
+            {twodays}
           </DueDate>
-        </RadioBtnConatiner>
-        <SubmittBtnConatiner>
+        </RadioBtnContainer>
+        <SubmittBtnContainer>
           <SubmitBtn type="submit">New todo</SubmitBtn>
-        </SubmittBtnConatiner>
+        </SubmittBtnContainer>
       </TodoForm>
     </Container>
   )
