+++ code/src/components/header/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Wrapper, Heading } from './header.style';
+
+const Header = () => {
+  return (
+    <Wrapper>
+      <Heading>TODO LIST </Heading>
+    </Wrapper>
+  )
+}
+
+export default Header
